add deciamls

+++ src/engine/engine.rs
@@ -1,5 +1,7 @@
 use std::collections::HashMap;
 
+use rust_decimal::Decimal;
+
 use super::orderbook::{Order, Orderbook};
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone)]
@@ -38,7 +40,7 @@ impl MatchingEngine {
     pub fn place_limit_order(
         &mut self,
         pair: TradingPair,
-        price: f64,
+        price: Decimal,
         order: Order,
     ) -> Result<(), String> {
         match self.orderbooks.get_mut(&pair) {
